created Admin Class and a couple of methods into that class for easier reading. moved a few methods into the customer class for easier reading. In Admin class added the AddAccount, BankCharges() and  bankInterest() methods.
Also moved the Make lodgements and withdraw funds functions from the Customer Menu into their own methods in the Customer Class. As well as any useful methods that would help with these new methods.

Added to comments to make easier to read and space out functionality in the Menu Class